requirements.txt
django>=2.1
python-decouple>=3.1
dj-database-url>=0.5.0
mysqlclient=1.3.12
django-separatedvaluesfield>=0.4
django-nested-admin>=3.1.2
django-colorful>=1.3.0
humanize>=0.5.1
django-admirarchy>=0.2.4
django-debug-toolbar>=1.10.1
django-filter>=2.0.0
djangorestframework>=3.8.2
drf-nested-routers>=0.91
markdown>=3.0.1
djangorestframework-camel-case>=0.2.0
pillow>=5.3.0
